More

+++ src/parser/flex/smt2_term_parser.cpp
@@ -68,7 +68,7 @@ size_t Smt2TermParser::parseIntegerNumeral() { return 0; }
 std::vector<DatatypeDecl> Smt2TermParser::parseDatatypeDef(
     bool isCo,
     const std::vector<std::string>& dnames,
-    const std::vector<int>& arities)
+    const std::vector<size_t>& arities)
 {
   std::vector<DatatypeDecl> dts;
   return dts;
